+++ guillotina/testing.py
@@ -11,6 +11,7 @@ import os
 TESTING_PORT = 55001
 
 TESTING_SETTINGS: Dict[str, Any] = {
+    "debug": True,
     "applications": ["guillotina.test_package"],
     "databases": {
         "db": {